+++ web-admin/src/features/projects/user-management/UserManagementGroupAvatar.svelte
