+++ web-admin/src/routes/[organization]/[project]/-/invite/+page.svelte
@@ -10,7 +10,7 @@
     getUserDomain,
     userDomainIsPublic,
   } from "@rilldata/web-admin/features/projects/user-invite/selectors";
-  import UserInviteForm from "@rilldata/web-admin/features/projects/user-invite/UserInviteForm.svelte";
+  import UserInviteForm from "@rilldata/web-admin/features/projects/user-management/UserInviteForm.svelte";
   import { Button } from "@rilldata/web-common/components/button";
   import Label from "@rilldata/web-common/components/forms/Label.svelte";
   import Switch from "@rilldata/web-common/components/forms/Switch.svelte";
